refactor(analytics): move bridge tracking to server-verified Iris flow
